Fix Error

+++ src/main/java/com/example/back/backend/application/dto/email/EMA0300AInput.java
@@ -13,6 +13,6 @@ public class EMA0300AInput {
     @NotNull
     String newAdminEmail;
     @NotNull
-    Long adminId;
+    int adminId;
 
 }
